Make the `ring` dependency optional
Make the dependency on `ring` truly optional.
It is still used by default through the `ring` feature,
but when disabling we don't depend on the `ring` dependency anymore.

This also means that the signature is now created by the configured
crypto provider as opposed to unconditionally by ring, which was the
case before. E.g.: Before, even without the `ring` dependency and with
the `aws-lc-rs` feature enabled signing of the `JWT` token was done
through the `ring` signatures.

When adding `gcp_auth` without any features enabled and without default
features, it will not compile anymore and instead throw a
`compile_error!`, which explains that we need at least one provider and
how to enable it.

Closes: #153
